+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - `--role` flag to `teams invite`
 - `teams join` using an invitation token
+- `create` cmd has search for product, plan and region
 
 ### Fixed